longterm checkout button rendering, just logic is off
