+++ content/modals/US/long_term_xo.json
@@ -37,7 +37,8 @@
             "Use autopay for your payments. It's easy!"
         ],
         "cta": {
-            "buttonText": "Continue with Pay Monthly",
+            "buttonTextEligible": "Continue with Pay Monthly",
+            "buttonText": "Back to Checkout",
             "link": "https://www.paypal.com/link/to/long-term-application"
         },
         "disclosure": "Payment options subject to credit approval. Estimated payment amounts may change based on shipping, taxes, or updates to your purchase. Offer availability depends on the merchant and purchase amount, and may not be available for subscriptions or recurring payments. You must be of legal age in your U.S. state of residence. [Lender disclosure][State licensing disclosures]",
