+++ demo/standalone-modal.html
@@ -10,7 +10,8 @@
         <button class="learn-more">Learn more</button>
         <script>
             const modal = window.paypal.MessagesModal({
-                account: 'DEV_US_MULTI'
+                account: 'DEV_US_MULTI',
+                channel: 'CHECKOUT'
             });
             document.querySelector('.learn-more').addEventListener('click', () => {
                 modal.show();
