+++ src/components/modal/v2/lib/index.js
@@ -3,3 +3,4 @@ export * from './providers';
 export * from './utils';
 export * from './logos';
 export * from './locale';
+export * from './getComputedVariables';
